This PR fixes the multi_add part of #5055.