@kmaehashi 2018-07-04
This issue has been introduced in #4679.