@kmaehashi 2018-07-04
I ran benchmark ([functions.math.basic_math.AddFunc.time_forward](https://github.com/chainer/chainer-benchmark/pull/25)) and confirmed performance regression fixed.

### this PR (with `inputs_all_ready` fixed)

```
=========== ========= ========= ==========
--                      count             
----------- ------------------------------
  backend       2         4         16    
=========== ========= ========= ==========
    gpu       1.95ms    6.56ms   28.26ms  
    cpu      26.11ms   45.05ms   156.65ms 
 cpu-ideep   10.06ms   17.91ms   57.19ms  
=========== ========= ========= ==========
```

### current master (3be74f5)

```
=========== ========= ========= ==========
--                      count             
----------- ------------------------------
  backend       2         4         16    
=========== ========= ========= ==========
    gpu       1.94ms    5.69ms   28.29ms  
    cpu      26.16ms   44.88ms   156.41ms 
 cpu-ideep   10.61ms   54.74ms   210.57ms 
=========== ========= ========= ==========
```

### 5.0.0a1

```
=========== ========= ========= ==========
--                      count             
----------- ------------------------------
  backend       2         4         16    
=========== ========= ========= ==========
    gpu       1.94ms    5.77ms   28.26ms  
    cpu      26.12ms   44.88ms   157.21ms 
 cpu-ideep   10.04ms   17.55ms   55.83ms  
=========== ========= ========= ==========
```